opt/slackbutt/requirements.txt
backports.ssl-match-hostname==3.5.0.1
peewee==2.8.5
psycopg2==2.6.2
pyowm==2.6.0
python-dateutil==2.6.0
PyYAML==3.12
requests==2.12.4
six==1.10.0
slackbot==0.4.1
slacker==0.9.30
sseclient==0.0.14
websocket-client==0.40.0
